Test for errors that should be thrown
When setting up a winston-loggly transport, tokena nd subdomain are
required and winston-loggly will throw errors if one of them is
missing. Test to make sure the expected error occurs when the
environment is not production (since we are going to supress the error
whe the environment _is_ production). WEB-2348